+++ app/decorators/residency_decorator.rb
@@ -20,8 +20,14 @@ class ResidencyDecorator < Draper::Decorator
             else
               lot_fees_paid? ? '' : '' # house-circle-check, house-circle-xmark
             end
+    css_classes = %w[fas icon]
+    css_classes << 'primary_residence' if primary_residence?
+    css_classes << 'second_home' unless primary_residence?
 
-    h.content_tag(:span, icon, class: 'fas', data: {tooltip: property.to_s})
+    tooltip = property.to_s
+    tooltip += ':2nd Home' unless primary_residence
+
+    h.content_tag(:span, icon, class: css_classes.join(' '), data: {tooltip: tooltip})
   end
 
   def resident_status_character
@@ -42,8 +48,11 @@ class ResidencyDecorator < Draper::Decorator
     end
   end
 
-  def resident_status_tag
-    h.content_tag(:span, resident_status_character, class: 'fas', data: {tooltip: resident_status_i18n})
+  def resident_status_tag(tooltip: resident_status_i18n)
+    css_classes = %w[fas icon]
+    css_classes << 'primary_residence' if primary_residence?
+    css_classes << 'second_home' unless primary_residence?
+    h.content_tag(:span, resident_status_character, class: css_classes.join(' '), data: {tooltip: tooltip})
   end
 
   def resident_status_i18n
@@ -73,6 +82,7 @@ class ResidencyDecorator < Draper::Decorator
     summary += "("
     summary += resident_status_i18n
     summary += ", minor" if is_minor?
+    summary += ", 2nd home" unless primary_residence?
     summary += ")"
     summary
   end