+++ db/migrate/20220525034325_add_primary_residence_to_residency.rb
@@ -1,0 +1,5 @@
+class AddPrimaryResidenceToResidency < ActiveRecord::Migration[7.0]
+  def change
+    add_column :residencies, :primary_residence, :boolean
+  end
+end