+++ db/migrate/20220525034325_add_residence_to_residency.rb
@@ -1,5 +1,0 @@
-class AddResidenceToResidency < ActiveRecord::Migration[7.0]
-  def change
-    add_column :residencies, :residence, :boolean
-  end
-end